+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Make sure `Generator` ignores deallocators on `const` values returned from adapters ([issue #317](https://github.com/bytedeco/javacpp/issues/317))
  * Accelerate `Loader.extractResource()` for directories already cached, also preventing failures ([issue #197](https://github.com/bytedeco/javacpp/issues/197))
  * Avoid `Parser` writing `allocateArray()` when single `int`, `long`, `float`, or `double` constructor already exists ([issue bytedeco/javacv#1224](https://github.com/bytedeco/javacv/issues/1224))
  * Expose all platform properties to process executed with `Builder.buildCommand` via environment variables, with names uppercase and all `.` replaced with `_`